+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Master (Unreleased)
 
+## 2.26.0 (2024-01-04)
+
 - Add new `RSpec/RedundantPredicateMatcher` cop. ([@ydah])
-- Add support for correcting "it will" (future tense) for `RSpec/ExampleWording`. ([@jdufresne])
 - Add new `RSpec/RemoveConst` cop. ([@swelther])
-- Ensure `PendingWithoutReason` can detect violations inside shared groups. ([@robinaugh])
+- Add support for correcting "it will" (future tense) for `RSpec/ExampleWording`. ([@jdufresne])
 - Add support for `symbol` style for `RSpec/SharedExamples`. ([@jessieay])
+- Ensure `PendingWithoutReason` can detect violations inside shared groups. ([@robinaugh])
 
 ## 2.25.0 (2023-10-27)